Fix SCAN Limit wire-width mismatch with the fitz broker (#33)
KvScanQuery.Limit was declared ulong? and written on the wire with
WriteU64 (8 bytes), but the fitz broker's SCAN decoder
(read_optional_limit) only reads a u32 (4 bytes) -- matching
MAX_SCAN_ITEMS and every other broker-side limit. The extra 4 bytes
shifted every subsequent field (reverse, optional start_exclusive),
which the broker's decoder correctly flagged as trailing data:

  Cntryl.Fitz.KvException: SCAN failed: Invalid request: Trailing
  data in SCAN payload

This meant every KvDirectory<T,TKey>.QueryAsync/ScanAllAsync call
against a real broker failed -- point reads/writes were unaffected,
only Scan. Root-caused by rebuilding the fitz broker from its own
current main and reproducing the identical error locally, confirming
this is a genuine field-width bug rather than a stale broker image.

Fixes it by changing KvScanQuery.Limit to uint?/WriteU32 to match the
broker's u32, updating the three (ulong) call sites in KvDirectory
that constructed it, and the two tests that asserted on the old type.

+++ src/Fitz.Extensions/KvDirectory.cs
@@ -178,9 +178,9 @@ public sealed class KvDirectory<T, TKey>
         var cursorKey = DecodeCursor(query.Cursor, fingerprint);
         ValidateCursorRange(cursorKey, rangeStart, rangeEnd);
         var scan = query.IsDescending
-            ? new KvScanQuery(rangeStart, cursorKey ?? rangeEnd, (ulong)(limit + 1), Reverse: true)
+            ? new KvScanQuery(rangeStart, cursorKey ?? rangeEnd, (uint)(limit + 1), Reverse: true)
             : new KvScanQuery(cursorKey is null ? rangeStart : After(cursorKey.Value.Span), rangeEnd,
-                (ulong)(limit + 1));
+                (uint)(limit + 1));
         await using var transaction = await client.BeginAsync(route, KvDurability.Async, KvMode.ReadOnly, ct)
             .ConfigureAwait(false);
         var matches = new List<KvPair>(limit + 1);
@@ -223,7 +223,7 @@ public sealed class KvDirectory<T, TKey>
         var cursorKey = DecodeCursor(cursor, fingerprint);
         ValidateCursorRange(cursorKey, rangeStart, rangeEnd);
         var scan = new KvScanQuery(
-            cursorKey is null ? rangeStart : After(cursorKey.Value.Span), rangeEnd, (ulong)(limit + 1));
+            cursorKey is null ? rangeStart : After(cursorKey.Value.Span), rangeEnd, (uint)(limit + 1));
         await using var transaction = await client.BeginAsync(route, KvDurability.Async, KvMode.ReadWrite, ct)
             .ConfigureAwait(false);
         var records = new List<KvPair>(limit + 1);
